Added --leak-temp-dir option to run_isolated.py for debugging isolates.
BUG=swarming:147
R=maruel@chromium.org

Review URL: https://codereview.appspot.com/132620043

+++ run_isolated.py
@@ -724,7 +724,7 @@ def process_command(command, out_dir):
   return filtered
 
 
-def run_tha_test(isolated_hash, storage, cache, extra_args):
+def run_tha_test(isolated_hash, storage, cache, leak_temp_dir, extra_args):
   """Downloads the dependencies in the cache, hardlinks them into a temporary
   directory and runs the executable from there.
 
@@ -741,6 +741,8 @@ def run_tha_test(isolated_hash, storage, cache, extra_args):
     cache: an isolateserver.LocalCache to keep from retrieving the same objects
            constantly by caching the objects retrieved. Can be on-disk or
            in-memory.
+    leak_temp_dir: if true, the temporary directory will be deliberately leaked
+                   for later examination.
     extra_args: optional arguments to add to the command stated in the .isolate
                 file.
   """
@@ -791,16 +793,21 @@ def run_tha_test(isolated_hash, storage, cache, extra_args):
 
   finally:
     try:
-      try:
-        if not rmtree(run_dir):
-          print >> sys.stderr, (
-              'Failed to delete the temporary directory, forcibly failing the\n'
-              'task because of it. No zombie process can outlive a successful\n'
-              'task run and still be marked as successful. Fix your stuff.')
-          result = result or 1
-      except OSError:
-        logging.warning('Leaking %s', run_dir)
-        result = 1
+      if leak_temp_dir:
+        logging.warning('Deliberately leaking %s for later examination',
+                        run_dir)
+      else:
+        try:
+          if not rmtree(run_dir):
+            print >> sys.stderr, (
+                'Failed to delete the temporary directory, forcibly failing\n'
+                'the task because of it. No zombie process can outlive a\n'
+                'successful task run and still be marked as successful.\n'
+                'Fix your stuff.')
+            result = result or 1
+        except OSError:
+          logging.warning('Leaking %s', run_dir)
+          result = 1
 
       # HACK(vadimsh): On Windows rmtree(run_dir) call above has
       # a synchronization effect: it finishes only when all task child processes
@@ -883,6 +890,14 @@ def main(args):
            'default=%default')
   parser.add_option_group(cache_group)
 
+  debug_group = optparse.OptionGroup(parser, 'Debugging')
+  debug_group.add_option(
+      '--leak-temp-dir',
+      action='store_true',
+      help='Deliberately leak isolate\'s temp dir for later examination '
+          '[default: %default]')
+  parser.add_option_group(debug_group)
+
   auth.add_auth_options(parser)
   options, args = parser.parse_args(args)
   auth.process_auth_options(parser, options)
@@ -908,7 +923,8 @@ def main(args):
     # Hashing schemes used by |storage| and |cache| MUST match.
     assert storage.hash_algo == cache.hash_algo
     return run_tha_test(
-        options.isolated or options.hash, storage, cache, args)
+        options.isolated or options.hash, storage, cache,
+        options.leak_temp_dir, args)
 
 
 if __name__ == '__main__':
